+++ src/ZPublisher/HTTPResponse.py
@@ -318,11 +318,7 @@ class HTTPBaseResponse(BaseResponse):
         `value` may be text or bytes. The default encoding of respective python
         version is used.
         """
-        cookies = self.cookies
-        if name in cookies:
-            cookie = cookies[name]
-        else:
-            cookie = cookies[name] = {}
+        cookie = {}
         for k, v in kw.items():
             k, v = convertCookieParameter(k, v)
             cookie[k] = v
@@ -330,6 +326,13 @@ class HTTPBaseResponse(BaseResponse):
         # RFC6265 makes quoting obsolete
         # cookie['quoted'] = quoted
         getCookieParamPolicy().check_consistency(name, cookie)
+
+        # update ``self.cookies`` only if there have been no exceptions
+        cookies = self.cookies
+        if name in cookies:
+            cookies[name].update(cookie)
+        else:
+            cookies[name] = cookie
 
     def appendCookie(self, name, value):
         """ Set an HTTP cookie.