+++ src/ZPublisher/cookie.py
@@ -250,25 +250,40 @@ registerCookieParameter("Domain", domain_converter)
 
 
 # ``Path``
-# we include "%" to handle the case that the path has already been quoted
-path_safe = compile(r"[a-zA-Z0-9_/%]*$").match
+path_safe = compile("["
+                    "-_.!~*'()"  # RFC 2396 section 2.3 (unreserved)
+                    "a-zA-Z0-9"  # letters and digits
+                    "/:@&=+$,"   # RFC 2396 section 3.3 (path reserved)
+                                 # excluding ``;``
+                    "]*$").match
 
 
 def path_converter(value):
-    """convert *value* to a path.
+    """convert *value* to a cookie path.
 
     The convertion is based on ``absolute_url_path``.
     If *value* lacks this method, it is assumed to be a string.
-    If the string contains ``%``, it is assumed that it is already
-    quoted.
+    If the string contains ``%``, it is used as is; otherwise,
+    it may be quoted by Python's ``urllib.parse.quote``.
+
+    **Note**:
+    According to RFC 6265 section 5.1.4 a cookie path match **does not**
+    unquote its arguments. It is therefore important to use
+    the same quoting algorithm for the URL and the cookie path.
+    If the cookie path contains only allowed characters
+    (RFC 2396 unreserved (section 2.3) and
+    RFC 2396 path special characters (section 3.3) excluding ``;``),
+    the value is taken verbatim; otherwise it is quoted
+    by Python's `urllib.parse.quote` (which
+    is used by `OFS.Traversable` as its ULR quoting). It quotes
+    all special characters apart from ``-_./``.
+    Quote yourself if this does not match your URL quoting.
     """
     ap = getattr(aq_base(value), "absolute_url_path", None)
     if ap is not None:
         return value.absolute_url_path()
-    if path_safe(value):
+    if "%" in value or path_safe(value):
         return value
-    if "%" in value:
-        raise ValueError("not path safe but apparently quoted")
     return quote(value)